Move OneLakeDataSource to Common
